requirements.txt
Pillow==10.4.0
requests==2.32.5
pyttsx3==2.99
SpeechRecognition==3.14.3
ttkbootstrap==1.14.2
PyAudio==0.2.14
charset_normalizer==3.4.3
idna==3.10
urllib3==2.5.0
certifi==2025.8.3
typing-extensions==4.15.0
standard-aifc==3.13.0
audioop-lts==0.2.2
standard-chunk==3.13.0
